feat: move use auth controller hook to design system
