+++ libs/design-system/package.json
@@ -1,39 +1,61 @@
 {
   "name": "@novu/design-system",
   "version": "0.20.0",
-  "repository": "https://github.com/novuhq/novu",
   "description": "",
-  "scripts": {
-    "start": "npm run build:watch",
-    "prebuild": "rimraf dist",
-    "lint": "eslint --ext .ts,.tsx src",
-    "build": "npm run build:cjs && npm run build:esm && npm run build:types",
-    "build:cjs": "cross-env node_modules/.bin/tsc -p tsconfig.json",
-    "build:esm": "cross-env node_modules/.bin/tsc -p tsconfig.esm.json",
-    "build:esm:watch": "cross-env node_modules/.bin/tsc -p tsconfig.esm.json -w --preserveWatchOutput",
-    "build:types": "tsc --declaration --emitDeclarationOnly --declarationMap --declarationDir dist/types -p tsconfig.json",
-    "storybook": "storybook dev -p 6006",
-    "build-storybook": "storybook build",
-    "cypress:install": "cypress install",
-    "cypress:open": "cross-env NODE_ENV=test cypress open",
-    "cypress:run": "cross-env NODE_OPTIONS=--max_old_space_size=4096 NODE_ENV=test cypress run --component"
-  },
-  "author": "",
+  "repository": "https://github.com/novuhq/novu",
   "license": "ISC",
+  "author": "",
+  "sideEffects": false,
   "main": "dist/cjs/index.js",
   "module": "dist/esm/index.js",
+  "types": "dist/types/index.d.ts",
   "files": [
     "dist/cjs",
     "dist/esm",
     "dist/types"
   ],
-  "sideEffects": false,
+  "scripts": {
+    "prebuild": "rimraf dist",
+    "build": "npm run build:cjs && npm run build:esm && npm run build:types",
+    "build-storybook": "storybook build",
+    "build:cjs": "cross-env node_modules/.bin/tsc -p tsconfig.json",
+    "build:esm": "cross-env node_modules/.bin/tsc -p tsconfig.esm.json",
+    "build:esm:watch": "cross-env node_modules/.bin/tsc -p tsconfig.esm.json -w --preserveWatchOutput",
+    "build:types": "tsc --declaration --emitDeclarationOnly --declarationMap --declarationDir dist/types -p tsconfig.json",
+    "cypress:install": "cypress install",
+    "cypress:open": "cross-env NODE_ENV=test cypress open",
+    "cypress:run": "cross-env NODE_OPTIONS=--max_old_space_size=4096 NODE_ENV=test cypress run --component",
+    "lint": "eslint --ext .ts,.tsx src",
+    "start": "npm run build:watch",
+    "storybook": "storybook dev -p 6006"
+  },
+  "dependencies": {
+    "@cypress/react": "^7.0.3",
+    "@cypress/webpack-preprocessor": "^6.0.0",
+    "@emotion/css": "^11.10.5",
+    "@emotion/react": "^11.7.1",
+    "@emotion/styled": "^11.6.0",
+    "@mantine/core": "^5.7.1",
+    "@mantine/hooks": "^5.7.1",
+    "@mantine/notifications": "^5.7.1",
+    "@novu/client": "^0.20.0",
+    "@novu/shared": "^0.20.0",
+    "@segment/analytics-next": "^1.48.0",
+    "@sentry/react": "^7.40.0",
+    "@tanstack/react-query": "^4.20.4",
+    "axios": "^1.3.3",
+    "jwt-decode": "^3.1.2",
+    "react-helmet-async": "^1.3.0",
+    "react-hook-form": "7.43.9",
+    "react-table": "^7.8.0",
+    "tslib": "^2.3.1"
+  },
   "devDependencies": {
     "@storybook/addon-actions": "^7.5.0",
     "@storybook/addon-docs": "^7.4.2",
-    "@storybook/theming": "^7.4.2",
     "@storybook/react": "^7.4.2",
     "@storybook/react-webpack5": "^7.4.2",
+    "@storybook/theming": "^7.4.2",
     "@types/node": "^18.11.12",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
@@ -58,21 +80,5 @@
     "react": ">=16.8.0",
     "react-dom": ">=16.8.0",
     "react-router-dom": "6.2.2"
-  },
-  "dependencies": {
-    "@cypress/react": "^7.0.3",
-    "@cypress/webpack-preprocessor": "^6.0.0",
-    "@emotion/css": "^11.10.5",
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@mantine/core": "^5.7.1",
-    "@mantine/hooks": "^5.7.1",
-    "@mantine/notifications": "^5.7.1",
-    "@novu/client": "^0.20.0",
-    "@novu/shared": "^0.20.0",
-    "react-hook-form": "7.43.9",
-    "react-table": "^7.8.0",
-    "react-helmet-async": "^1.3.0",
-    "tslib": "^2.3.1"
   }
 }
