+++ libs/design-system/src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useLocalThemePreference';
 export * from './useDataRef';
 export * from './useKeyDown';
+export * from './useAuthController';
