+++ libs/design-system/src/utils/segment.ts
